# backend/swagger.yaml
openapi: 3.0.0
info:
  title: API админ-панели
  version: 1.0.0
  description: API для управления товарами в интернет-магазине
servers:
  - url: http://localhost:4000
    description: Локальный сервер
paths:
  /api/products:
    get:
      summary: Получить список всех товаров
      responses:
        '200':
          description: Список товаров
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Product'
    post:
      summary: Добавить новый товар
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Product'
      responses:
        '201':
          description: Товар успешно добавлен
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Product'
  /api/products/{id}:
    get:
      summary: Получить товар по ID
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        '200':
          description: Товар найден
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Product'
        '404':
          description: Товар не найден
    put:
      summary: Обновить товар по ID
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Product'
      responses:
        '200':
          description: Товар успешно обновлен
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Product'
        '404':
          description: Товар не найден
    delete:
      summary: Удалить товар по ID
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        '200':
          description: Товар успешно удален
        '404':
          description: Товар не найден
components:
  schemas:
    Product:
      type: object
      properties:
        id:
          type: integer
          description: Уникальный идентификатор товара
        name:
          type: string
          description: Название товара
        price:
          type: number
          description: Цена товара
        description:
          type: string
          description: Описание товара
        categories:
          type: array
          items:
            type: string
          description: Категории товара
      required:
        - name
        - price
        - description
        - categories